CH4: add rest controller to handle the creation of orders & add implementation of strategies for finding the location
Service selects a strategy in order to find the location based on a
configuration, after that it creates an order and saves it to the
database. It should also update the stock. The rest controller uses this
service, it has a POST mapping to accept a json request body to create
the order.